Keep track of rounding errors in REL_ send code, so that small values don't get lost, leads to better mouse emulation

+++ src/axis_event.cpp
@@ -246,7 +246,7 @@ RelAxisEventHandler::send(UInput& uinput, int value)
   else
     value_f = static_cast<float>(value) / static_cast<float>(m_max);
 
-  int v = static_cast<int>(m_value * value_f);
+  float v = m_value * value_f;
 
   if (v == 0)
     uinput.send_rel_repetitive(m_code, v, -1);
